create components for login and registration
