+++ frontend/components/user/Login.vue
@@ -1,6 +1,31 @@
-import { UserResponseObject, UserObject } from '../stateless/interfaces/response-objects';
 <template>
-
+  <form class="card-form" @submit.prevent="handleLogin">
+    <fieldset class="field-group">
+      <span v-text="userLoginError"></span>
+      <div>
+        <UserFormInput
+          :label="identifierLabel"
+          :index="0"
+          :isPassword="false"
+          :placeholder="identifierLabel"
+          v-model:inputValue="identifier"
+        />
+        <UserFormInput
+          :label="passwordLabel"
+          :index="1"
+          :isPassword="true"
+          :placeholder="passwordLabel"
+          v-model:inputValue="password"
+        />
+      </div>
+    </fieldset>
+    <button
+      type="submit"
+      class="pure-button pure-input-1-2 pure-button-primary"
+    >
+      Login
+    </button>
+  </form>
 </template>
 
 <script setup lang="ts">
@@ -16,7 +41,6 @@ const identifierLabel = "Username or Email";
 const passwordLabel = "Password";
 const identifier = ref("");
 const password = ref("");
-
 
 const userStore = getUserStore();
 const router = useRouter();
